Extract resize callbacks to separate file
