+++ js/bundles/modules/core.js
@@ -81,7 +81,7 @@ DevExpress.utils.date = require("../../core/utils/date");
 DevExpress.utils.browser = require("../../core/utils/browser");
 DevExpress.utils.inflector = require("../../core/utils/inflector");
 DevExpress.utils.iterator = require("../../core/utils/iterator");
-DevExpress.utils.resizeCallbacks = windowUtils.resizeCallbacks;
+DevExpress.utils.resizeCallbacks = require("../../core/utils/resize_callbacks");
 DevExpress.utils.console = require("../../core/utils/console");
 DevExpress.utils.string = require("../../core/utils/string");
 DevExpress.utils.support = require("../../core/utils/support");
