+++ js/core/devices.js
@@ -10,7 +10,7 @@ var $ = require("../core/renderer"),
     Class = require("./class"),
     errors = require("./errors"),
     Callbacks = require("./utils/callbacks"),
-    resizeCallbacks = windowUtils.resizeCallbacks,
+    resizeCallbacks = require("./utils/resize_callbacks"),
     EventsMixin = require("./events_mixin"),
     SessionStorage = require("./utils/storage").sessionStorage,
     viewPort = require("./utils/view_port"),
