+++ js/core/dom_component.js
@@ -6,8 +6,7 @@ var $ = require("../core/renderer"),
     config = require("./config"),
     errors = require("./errors"),
     getPublicElement = require("../core/utils/dom").getPublicElement,
-    windowUtils = require("../core/utils/window"),
-    windowResizeCallbacks = windowUtils.resizeCallbacks,
+    windowResizeCallbacks = require("../core/utils/resize_callbacks"),
     commonUtils = require("./utils/common"),
     each = require("./utils/iterator").each,
     typeUtils = require("./utils/type"),
