+++ js/core/utils/window.js
@@ -3,9 +3,7 @@
 /* global window */
 
 var domAdapter = require("../dom_adapter"),
-    callOnce = require("./call_once"),
-    readyCallbacks = require("./ready_callbacks"),
-    Callbacks = require("./callbacks");
+    callOnce = require("./call_once");
 
 var hasWindow = function() {
     return typeof window !== "undefined";
@@ -18,76 +16,6 @@ var getWindow = function() {
 var hasProperty = function(prop) {
     return hasWindow() && prop in window;
 };
-
-var resizeCallbacks = (function() {
-    var prevSize,
-        callbacks = Callbacks(),
-        originalCallbacksAdd = callbacks.add,
-        originalCallbacksRemove = callbacks.remove;
-
-    if(!hasWindow()) {
-        return callbacks;
-    }
-
-    var formatSize = function() {
-        var documentElement = domAdapter.getDocumentElement();
-        return {
-            width: documentElement.clientWidth,
-            height: documentElement.clientHeight
-        };
-    };
-
-    var handleResize = function() {
-        var now = formatSize();
-        if(now.width === prevSize.width && now.height === prevSize.height) {
-            return;
-        }
-
-        var changedDimension;
-        if(now.width === prevSize.width) {
-            changedDimension = 'height';
-        }
-        if(now.height === prevSize.height) {
-            changedDimension = 'width';
-        }
-
-        prevSize = now;
-
-        callbacks.fire(changedDimension);
-    };
-
-    var setPrevSize = callOnce(function() {
-        prevSize = formatSize();
-    });
-
-    var removeListener;
-
-    callbacks.add = function() {
-        var result = originalCallbacksAdd.apply(callbacks, arguments);
-
-        setPrevSize();
-
-        readyCallbacks.add(function() {
-            if(!removeListener && callbacks.has()) {
-                removeListener = domAdapter.listen(getWindow(), "resize", handleResize);
-            }
-        });
-
-        return result;
-    };
-
-    callbacks.remove = function() {
-        var result = originalCallbacksRemove.apply(callbacks, arguments);
-
-        if(!callbacks.has() && removeListener) {
-            removeListener();
-            removeListener = undefined;
-        }
-        return result;
-    };
-
-    return callbacks;
-})();
 
 var defaultScreenFactorFunc = function(width) {
     if(width < 768) {
@@ -126,7 +54,6 @@ var getNavigator = function() {
     };
 };
 
-exports.resizeCallbacks = resizeCallbacks;
 exports.defaultScreenFactorFunc = defaultScreenFactorFunc;
 exports.getCurrentScreenFactor = getCurrentScreenFactor;
 exports.beforeActivateExists = beforeActivateExists;
