+++ js/mobile/init_mobile_viewport/init_mobile_viewport.js
@@ -6,7 +6,7 @@ var $ = require("../../core/renderer"),
     window = windowUtils.getWindow(),
     eventsEngine = require("../../events/core/events_engine"),
     extend = require("../../core/utils/extend").extend,
-    resizeCallbacks = windowUtils.resizeCallbacks,
+    resizeCallbacks = require("../../core/utils/resize_callbacks"),
     support = require("../../core/utils/support"),
     styleUtils = require("../../core/utils/style"),
     devices = require("../../core/devices");
