+++ js/viz/core/base_widget.js
@@ -3,8 +3,7 @@
 var noop = require("../../core/utils/common").noop,
     each = require("../../core/utils/iterator").each,
     version = require("../../core/version"),
-    windowUtils = require("../../core/utils/window"),
-    _windowResizeCallbacks = windowUtils.resizeCallbacks,
+    _windowResizeCallbacks = require("../../core/utils/resize_callbacks"),
     _stringFormat = require("../../core/utils/string").format,
     _isObject = require("../../core/utils/type").isObject,
     extend = require("../../core/utils/extend").extend,
