+++ testing/tests/DevExpress.core/domComponent.tests.js
@@ -4,7 +4,7 @@ var $ = require("jquery"),
     noop = require("core/utils/common").noop,
     registerComponent = require("core/component_registrator"),
     config = require("core/config"),
-    resizeCallbacks = require("core/utils/window").resizeCallbacks,
+    resizeCallbacks = require("core/utils/resize_callbacks"),
     devices = require("core/devices"),
     DOMComponent = require("core/dom_component"),
     domUtils = require("core/utils/dom"),
