+++ testing/tests/DevExpress.ui.widgets.dataGrid/dataGrid.tests.js
@@ -54,7 +54,7 @@ var DataGrid = require("ui/data_grid/ui.data_grid");
 var $ = require("jquery"),
     Class = require("core/class"),
     ODataUtils = require("data/odata/utils"),
-    resizeCallbacks = require("core/utils/window").resizeCallbacks,
+    resizeCallbacks = require("core/utils/resize_callbacks"),
     logger = require("core/utils/console").logger,
     errors = require("ui/widget/ui.errors"),
     commonUtils = require("core/utils/common"),
