+++ testing/tests/DevExpress.ui.widgets.scheduler/integration.appointmentTooltip.tests.js
@@ -16,7 +16,7 @@ require("generic_light.css!");
 var $ = require("jquery"),
     Tooltip = require("ui/tooltip"),
     tooltip = require("ui/tooltip/ui.tooltip"),
-    resizeCallbacks = require("core/utils/window").resizeCallbacks,
+    resizeCallbacks = require("core/utils/resize_callbacks"),
     fx = require("animation/fx"),
     dateLocalization = require("localization/date"),
     messageLocalization = require("localization/message"),
