+++ testing/tests/DevExpress.ui.widgets.scheduler/workSpace.tests.js
@@ -10,7 +10,7 @@ var $ = require("jquery"),
     SchedulerWorkSpace = require("ui/scheduler/ui.scheduler.work_space"),
     SchedulerResourcesManager = require("ui/scheduler/ui.scheduler.resource_manager"),
     domUtils = require("core/utils/dom"),
-    resizeCallbacks = require("core/utils/window").resizeCallbacks,
+    resizeCallbacks = require("core/utils/resize_callbacks"),
     dateUtils = require("core/utils/date"),
     dateLocalization = require("localization/date"),
     dragEvents = require("events/drag");
