+++ testing/tests/DevExpress.ui.widgets/gallery.tests.js
@@ -6,7 +6,7 @@ var $ = require("jquery"),
     ArrayStore = require("data/array_store"),
     fx = require("animation/fx"),
     animationFrame = require("animation/frame"),
-    resizeCallbacks = require("core/utils/window").resizeCallbacks,
+    resizeCallbacks = require("core/utils/resize_callbacks"),
     isRenderer = require("core/utils/type").isRenderer,
     config = require("core/config"),
     executeAsyncMock = require("../../helpers/executeAsyncMock.js"),
