+++ testing/tests/DevExpress.ui.widgets/menu.tests.js
@@ -7,7 +7,7 @@ var $ = require("jquery"),
     isRenderer = require("core/utils/type").isRenderer,
     config = require("core/config"),
     Submenu = require("ui/menu/ui.submenu"),
-    resizeCallbacks = require("core/utils/window").resizeCallbacks,
+    resizeCallbacks = require("core/utils/resize_callbacks"),
     Menu = require("ui/menu/ui.menu"),
     keyboardMock = require("../../helpers/keyboardMock.js"),
     fixtures = require("../../helpers/positionFixtures.js");
