+++ testing/tests/DevExpress.ui.widgets/overlay.tests.js
@@ -9,7 +9,7 @@ var $ = require("jquery"),
     hideTopOverlayCallback = require("mobile/hide_top_overlay").hideCallback,
     positionUtils = require("animation/position"),
     domUtils = require("core/utils/dom"),
-    resizeCallbacks = require("core/utils/window").resizeCallbacks,
+    resizeCallbacks = require("core/utils/resize_callbacks"),
     devices = require("core/devices"),
     Template = require("ui/widget/jquery.template"),
     Overlay = require("ui/overlay"),
