+++ testing/tests/DevExpress.ui.widgets/scrollable.tests.js
@@ -9,7 +9,7 @@ var $ = require("jquery"),
     animationFrame = require("animation/frame"),
     domUtils = require("core/utils/dom"),
     initMobileViewport = require("mobile/init_mobile_viewport"),
-    resizeCallbacks = require("core/utils/window").resizeCallbacks,
+    resizeCallbacks = require("core/utils/resize_callbacks"),
     devices = require("core/devices"),
     Scrollable = require("ui/scroll_view/ui.scrollable"),
     simulatedStrategy = require("ui/scroll_view/ui.scrollable.simulated"),
