+++ testing/tests/DevExpress.ui.widgets/slideOutView.tests.js
@@ -4,7 +4,7 @@ var $ = require("jquery"),
     fx = require("animation/fx"),
     translator = require("animation/translator"),
     hideTopOverlayCallback = require("mobile/hide_top_overlay").hideCallback,
-    resizeCallbacks = require("core/utils/window").resizeCallbacks,
+    resizeCallbacks = require("core/utils/resize_callbacks"),
     config = require("core/config"),
     typeUtils = require("core/utils/type"),
     animation = require("ui/slide_out_view").animation,
