+++ testing/tests/DevExpress.viz.gauges/common.tests.js
@@ -6,7 +6,7 @@ var $ = require("jquery"),
     noop = require("core/utils/common").noop,
     Class = require("core/class"),
     registerComponent = require("core/component_registrator"),
-    resizeCallbacks = require("core/utils/window").resizeCallbacks,
+    resizeCallbacks = require("core/utils/resize_callbacks"),
     objectUtils = require("core/utils/object"),
     vizMocks = require("../../helpers/vizMocks.js"),
     dxGauge = require("viz/gauges/common").dxGauge,
